fixed VoteForm loading + refactored
