+++ apps/zupoll-client/components/main/ErrorOverlay.tsx
@@ -1,5 +1,7 @@
 import * as React from "react";
 import { Overlay } from "./Overlay";
+import { ZupollError } from "../../src/types";
+
 
 export function ErrorOverlay({
   error,
@@ -24,13 +26,4 @@ export function ErrorOverlay({
       <button onClick={onClose}>Close</button>
     </Overlay>
   );
-}
-
-export interface ZupollError {
-  /** Big title, should be under 40 chars */
-  title: string;
-  /** Useful explanation, avoid "Something went wrong." */
-  message: string | React.ReactNode;
-  /** Optional stacktrace. */
-  stack?: string;
 }
